add interpret tests and complete api tests
